Changed Alumni class, zip code matching, handling confidential data, reading in student
In Alumni class, I added the statedPriority ArrayList. The distance between each alumnus and student pair is considered instead of just the first two digits of each zip code. Confidential files are now placed in the confidential_data folder and stated in the gitignore file. Reading in students is now more robust to missing values.